+++ src/render.js
@@ -35,10 +35,19 @@ const createElement = (template) => {
 };
 
 const replace = (newElement, oldElement) => {
+
+  if (newElement === null || oldElement === null) {
+    throw new Error('Can\'t replace unexisting elements');
+  }
+
   const newChild = newElement instanceof AbstractView ? newElement.element : newElement;
   const oldChild = oldElement instanceof AbstractView ? oldElement.element : oldElement;
 
   const parent = oldChild.parentElement;
+
+  if (parent === null) {
+    throw new Error('Parent element doesn\'t exist');
+  }
 
   parent.replaceChild(newChild, oldChild);
 };
