+++ src/view/top-route-info-view.js
